server/requirements.txt
torch==2.0.0
ffmpeg-python==0.2.0
openai-whisper==20230314
